ViewUtils moved to views package
